Move common load methods into cldr/load/default
`cldr/load/webpack` no longer includes the values imported from `cldr/load`
in webpack builds. Separating the common methods between the two into a
shared module circumvents this issue. To preserve the existing API, exports
from `cldr/load/default` are re-exported by both `cldr/load` and
`cldr/load/webpack`.
