+++ tests/unit/cldr/load.ts
@@ -6,6 +6,12 @@ import loadCldrData, {
 	reset,
 	supplementalPackages
 } from '../../../src/cldr/load';
+import {
+	isLoaded as utilIsLoaded,
+	mainPackages as utilMainPackages,
+	reset as utilReset,
+	supplementalPackages as utilSupplementalPackages
+} from '../../../src/cldr/load/default';
 
 registerSuite({
 	name: 'cldr/load',
@@ -14,71 +20,11 @@ registerSuite({
 		reset();
 	},
 
-	mainPackages() {
-		assert.isTrue(Object.isFrozen(mainPackages), 'Should be frozen.');
-		assert.sameMembers(mainPackages as any[], [
-			'dates/calendars/gregorian',
-			'dates/fields',
-			'dates/timeZoneNames',
-			'numbers',
-			'numbers/currencies',
-			'units'
-		]);
-	},
-
-	supplementalPackages() {
-		assert.isTrue(Object.isFrozen(supplementalPackages), 'Should be frozen.');
-		assert.sameMembers(supplementalPackages as any[], [
-			'currencyData',
-			'likelySubtags',
-			'numberingSystems',
-			'plurals-type-cardinal',
-			'plurals-type-ordinal',
-			'timeData',
-			'weekData'
-		]);
-	},
-
-	isLoaded: {
-		'with an unloaded package'() {
-			assert.isFalse(isLoaded('supplemental', 'likelySubtags'));
-			assert.isFalse(isLoaded('main', 'en'));
-		},
-
-		'with loaded pacakges'() {
-			loadCldrData({
-				main: {
-					zh: {
-						numbers: {}
-					}
-				},
-
-				supplemental: {
-					likelySubtags: {}
-				}
-			});
-
-			assert.isTrue(isLoaded('main'));
-			assert.isTrue(isLoaded('supplemental'));
-			assert.isTrue(isLoaded('main', 'zh'));
-			assert.isTrue(isLoaded('main', 'zh-MO'));
-			assert.isTrue(isLoaded('main', 'zh', 'numbers'));
-			assert.isTrue(isLoaded('supplemental', 'likelySubtags'));
-		},
-
-		'with unknown packages'() {
-			loadCldrData({
-				main: {
-					arbitrary: {}
-				},
-				supplemental: {
-					arbitrary: {}
-				}
-			});
-
-			assert.isFalse(isLoaded('main', 'arbitrary'), 'Unknown locale packages are ignored.');
-			assert.isFalse(isLoaded('supplemental', 'arbitrary'), 'Unknown supplemental packages are ignored.');
-		}
+	api() {
+		assert.strictEqual(isLoaded, utilIsLoaded, 'isLoaded should be re-exported');
+		assert.strictEqual(mainPackages, utilMainPackages, 'mainPackages should be re-exported');
+		assert.strictEqual(reset, utilReset, 'reset should be re-exported');
+		assert.strictEqual(supplementalPackages, utilSupplementalPackages, 'supplementalPackages should be re-exported');
 	},
 
 	loadCldrData: {
@@ -100,51 +46,6 @@ registerSuite({
 			}).then(() => {
 				assert.isTrue(isLoaded('supplemental', 'likelySubtags'));
 			});
-		}
-	},
-
-	reset: {
-		beforeEach() {
-			loadCldrData({
-				main: {
-					zh: {
-						numbers: {}
-					}
-				},
-
-				supplemental: {
-					likelySubtags: {}
-				}
-			});
-		},
-
-		'main only'() {
-			reset('main');
-
-			assert.isFalse(isLoaded('main', 'zh'), '"main" data should be cleared.');
-			assert.isFalse(isLoaded('main', 'zh-MO'), '"main" data should be cleared.');
-			assert.isFalse(isLoaded('main', 'zh', 'numbers'), '"main" data should be cleared.');
-
-			assert.isTrue(isLoaded('supplemental', 'likelySubtags'), '"supplemental" data should not be cleared.');
-		},
-
-		'supplemental only'() {
-			reset('supplemental');
-
-			assert.isTrue(isLoaded('main', 'zh'));
-			assert.isTrue(isLoaded('main', 'zh-MO'));
-			assert.isTrue(isLoaded('main', 'zh', 'numbers'));
-
-			assert.isFalse(isLoaded('supplemental', 'likelySubtags'), '"supplemental" data should be cleared.');
-		},
-
-		'both main and supplmental'() {
-			reset();
-
-			assert.isFalse(isLoaded('main', 'zh'));
-			assert.isFalse(isLoaded('main', 'zh-MO'));
-			assert.isFalse(isLoaded('main', 'zh', 'numbers'));
-			assert.isFalse(isLoaded('supplemental', 'likelySubtags'));
 		}
 	}
 });
